refactor(pt): update config import router
