+++ packages/politics-tracker/components/landing/election-2024/fact-check-group/president-factcheck/index.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import CustomSelect from '~/components/landing/election-2024/fact-check-group/president-factcheck/custom-select'
 import FactCheckItem from '~/components/landing/election-2024/fact-check-group/president-factcheck/factcheck-item'
 import { FactCheckPresident } from '~/components/landing/react-context/landing-2024-context'
-import { prefixOfJSONForLanding2024 } from '~/constants/config'
+import { prefixOfJSONForLanding2024 } from '~/constants/environment-variables'
 import { checkboxLabels } from '~/constants/landing'
 import { defaultFactCheckJSON } from '~/constants/landing'
 import type { PoliticCategory } from '~/types/politics-detail'
