+++ packages/politics-tracker/pages/election.tsx
@@ -8,8 +8,9 @@ import type { GetServerSideProps } from 'next'
 import CustomHead, { type HeadProps } from '~/components/custom-head'
 import DefaultLayout from '~/components/layout/default'
 import Nav, { type LinkMember, NavProps } from '~/components/nav'
-import { cmsApiUrl, env } from '~/constants/config'
+import { cmsApiUrl } from '~/constants/config'
 import { districtsMapping, electionTypesMapping } from '~/constants/election'
+import { env } from '~/constants/environment-variables'
 import GetElection from '~/graphql/query/election/get-election.graphql'
 import GetElectionHistoryOfArea from '~/graphql/query/election/get-election-history-of-area.graphql'
 import type {
