+++ packages/politics-tracker/pages/server-sitemaps/landing-and-elections.xml/index.tsx
@@ -6,7 +6,8 @@ import { print } from 'graphql'
 import { GetServerSideProps } from 'next'
 import { getServerSideSitemap } from 'next-sitemap'
 
-import { cmsApiUrl, siteUrl, urlOfJsonForlandingPage } from '~/constants/config'
+import { cmsApiUrl, urlOfJsonForlandingPage } from '~/constants/config'
+import { siteUrl } from '~/constants/environment-variables'
 import GetElections from '~/graphql/query/sitemap/get-elections.graphql'
 import { GenericGQLData, RawElection } from '~/types/common'
 import { fireGqlRequest } from '~/utils/utils'
