+++ packages/politics-tracker/pages/server-sitemaps/person.xml/index.tsx
@@ -5,7 +5,8 @@ import { print } from 'graphql'
 import { GetServerSideProps } from 'next'
 import { getServerSideSitemap, ISitemapField } from 'next-sitemap'
 
-import { cmsApiUrl, siteUrl } from '~/constants/config'
+import { cmsApiUrl } from '~/constants/config'
+import { siteUrl } from '~/constants/environment-variables'
 import GetPeople from '~/graphql/query/sitemap/get-people.graphql'
 import { GenericGQLData, RawPerson } from '~/types/common'
 import { fireGqlRequest } from '~/utils/utils'
