+++ packages/politics-tracker/pages/server-sitemaps/politics-detail.xml/index.tsx
@@ -5,7 +5,8 @@ import { print } from 'graphql'
 import { GetServerSideProps } from 'next'
 import { getServerSideSitemap } from 'next-sitemap'
 
-import { cmsApiUrl, siteUrl } from '~/constants/config'
+import { cmsApiUrl } from '~/constants/config'
+import { siteUrl } from '~/constants/environment-variables'
 import GetPolitics from '~/graphql/query/sitemap/get-politics.graphql'
 import { GenericGQLData, RawPolitic } from '~/types/common'
 import { fireGqlRequest } from '~/utils/utils'
