+++ packages/politics-tracker/utils/analytics.ts
@@ -1,14 +1,12 @@
 import ReactGA from 'react-ga'
 
-import { gaTrackingId } from '~/constants/config'
+import { gaTrackingId } from '~/constants/environment-variables'
 
 export const initGA = () => {
-  // console.log('GA init')
   ReactGA.initialize(gaTrackingId)
 }
 
 export const logPageView = () => {
-  // console.log(`Logging pageview for ${window.location.pathname}`)
   ReactGA.set({ page: window.location.pathname })
   ReactGA.pageview(window.location.pathname)
 }
